Remove AccountExpires flag
If the internet of a machine stops working, then the password rotation also stops. This causes the previously created account to expire and no longer be functional. Thus, you can't login as an administrator. By removing the AccountExpires flag, the last admin account that synced to RMM will still work.